+++ src/main/java/gs/mclo/api/APIException.java
@@ -1,8 +1,11 @@
 package gs.mclo.api;
 
+import org.jetbrains.annotations.ApiStatus;
+
 import java.io.IOException;
 
-public class APIException extends IOException {
+public final class APIException extends IOException {
+    @ApiStatus.Internal
     public APIException(String message) {
         super(message);
     }