+++ src/main/java/gs/mclo/api/MclogsClient.java
@@ -3,6 +3,7 @@ package gs.mclo.api;
 import com.google.gson.Gson;
 import gs.mclo.api.response.*;
 import gs.mclo.api.util.JsonBodyHandler;
+import gs.mclo.api.util.Util;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;