+++ src/main/java/gs/mclo/api/response/insights/Problem.java
@@ -1,11 +1,15 @@
 package gs.mclo.api.response.insights;
 
 @SuppressWarnings("NotNullFieldNotInitialized")
-public class Problem extends Insight {
-    protected Solution[] solutions;
+public final class Problem extends Insight {
+    private Solution[] solutions;
+
+    private Problem() {
+    }
 
     /**
      * Get the solutions for this problem
+     *
      * @return possible solutions for this problem
      */
     public Solution[] getSolutions() {